Fixed updating and adding new students and added validation in backend

+++ src/main/java/com/rasmusrim/restapidemo/repositories/AbsenceRepositoryCustom.java
@@ -9,6 +9,6 @@ import java.util.List;
 @Repository
 public interface AbsenceRepositoryCustom {
     List<AbsenceEntry> getAbsenceEntriesByMonth(LocalDate date);
-    AbsenceEntry getAbsenceEntryByDateAndStudent(LocalDate date, int studentId);
+    AbsenceEntry getAbsenceEntryByDateAndStudent(LocalDate date, long studentId);
 
 }
